+++ src/ipc/StartListening.h
@@ -13,7 +13,7 @@
 
 #include "async/AsyncCall.h"
 #include "base/forward.h"
-#include "base/Subscription.h"
+#include "async/Subscription.h"
 #include "comm/forward.h"
 #include "ip/forward.h"
 #include "ipc/FdNotes.h"
