+++ src/servers/FtpServer.cc
@@ -13,7 +13,7 @@
 #include "base/CharacterSet.h"
 #include "base/Raw.h"
 #include "base/RefCount.h"
-#include "base/Subscription.h"
+#include "async/Subscription.h"
 #include "client_side_reply.h"
 #include "client_side_request.h"
 #include "clientStream.h"
